+++ src/main/java/com/bud/feature/profiles/GronkhProfile.java
@@ -6,6 +6,7 @@ import com.bud.core.sound.GronkhSound;
 import com.bud.core.sound.IBudSound;
 import com.bud.core.types.BudType;
 import com.bud.core.types.DayOfWeek;
+import com.bud.core.types.Pronoun;
 import com.bud.feature.LLMPromptManager;
 import com.bud.llm.messages.BudMessage;
 import com.bud.llm.profiles.IBudProfile;
@@ -61,5 +62,11 @@ public class GronkhProfile implements IBudProfile {
 	@Override
 	public DayOfWeek getFavoriteDay() {
 		return DayOfWeek.FRIDAY;
+	}
+
+	@Nonnull
+	@Override
+	public Pronoun getPronoun() {
+		return Pronoun.HE;
 	}
 }